+++ skore/src/skore/_sklearn/_cross_validation/metrics_accessor.py
@@ -15,10 +15,10 @@ from skore._sklearn._plot import (
     PredictionErrorDisplay,
     RocCurveDisplay,
 )
+from skore._sklearn._plot.metrics.metrics_summary_display import metric_score_to_rows
 from skore._sklearn.types import Aggregate, MetricLike
 from skore._utils._accessor import _check_estimator_report_has_method
 from skore._utils._fixes import _validate_joblib_parallel_params
-from skore._utils._metric_rows import metric_score_to_rows
 from skore._utils._parallel import delayed
 from skore._utils._progress_bar import track
 
@@ -233,7 +233,7 @@ class _MetricsAccessor(_BaseAccessor[CrossValidationReport], DirNamesMixin):
         reports = self._parent.estimator_reports_
         metric = reports[0]._metric_registry[metric_name]
 
-        rows: list[dict] = []
+        rows: list = []
         for split_idx, report in enumerate(reports):
             score = getattr(report.metrics, metric_name)(
                 data_source=data_source, **kwargs
